requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
firebase-admin==6.2.0
python-dotenv==1.0.0
Flask-Migrate
beautifulsoup4
groq
gunicorn
psycopg2-binary
